requirements.txt
streamlit
pandas
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
fpdf2 
qrcode
Pillow
ollama
python-docx
streamlit-drawable-canvas
ollama
python-docx
python-dotenv